Deleted Alarm.activity. Everything is moved to main_activity now. Implemented UI for adding a new alarm. Switch method not implemented.
