+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Alarm"></activity>
         <activity android:name=".Settings" />
         <activity android:name=".MainActivity">
             <intent-filter>
