--- app/src/main/java/com/example/alarmapk/Alarm.java
@@ -1,88 +1,0 @@
-package com.example.alarmapk;
-
-import androidx.annotation.RequiresApi;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.DialogFragment;
-
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.app.TimePickerDialog;
-import android.content.Context;
-import android.content.Intent;
-import android.os.Build;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-import android.widget.TimePicker;
-
-import java.text.DateFormat;
-import java.util.Calendar;
-
-public class Alarm extends AppCompatActivity implements TimePickerDialog.OnTimeSetListener {
-
-    Button setAlarmButton, finishButton;
-    TextView textView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_alarm);
-        setAlarmButton = findViewById(R.id.setAlarm);
-        textView = findViewById(R.id.textView);
-        finishButton = findViewById(R.id.finishButton);
-        //otvaranje timera za alarm
-        setAlarmButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                DialogFragment timePicker = new TimePickerFragment();
-                timePicker.show(getSupportFragmentManager(), "time picker");
-
-            }
-        });
-
-        finishButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                finish(); //closing intent
-            }
-        });
-
-
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-    @Override
-    public void onTimeSet(TimePicker timePicker, int i, int i1) {
-        Calendar c = Calendar.getInstance();
-        c.set(Calendar.HOUR_OF_DAY, i);
-        c.set(Calendar.MINUTE, i1);
-        c.set(Calendar.SECOND, 0);
-        updateTimeText(c);
-        startAlarm(c);
-        System.out.println("ALARM JE POSTAVLJEN");
-
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-    private void startAlarm(Calendar c) {
-        AlarmManager alarmManager = (AlarmManager) getSystemService(Context.ALARM_SERVICE);
-        Intent intent = new Intent(this, MyBroadcastReceiver.class);
-        PendingIntent pendingIntent = PendingIntent.getBroadcast(this, 1, intent, 0);
-
-        if (c.before(Calendar.getInstance())) {
-            c.add(Calendar.DATE, 1);  // If time in past :D date tomorow
-        }
-
-        alarmManager.setExact(AlarmManager.RTC_WAKEUP, c.getTimeInMillis(), pendingIntent);
-    }
-
-    private void updateTimeText(Calendar c) { //update na ekranu kad je alarm
-        String timeText = "Alarm set for: ";
-        timeText += DateFormat.getTimeInstance(DateFormat.SHORT).format(c.getTime());
-
-        textView.setText(timeText);
-    }
-
-
-}
