+++ app/src/main/res/layout/activity_main.xml
@@ -52,105 +52,12 @@
         app:layout_constraintVertical_bias="0.896">
 
         <LinearLayout
+            android:id="@+id/linearlay"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-
-            <TextView
-                android:id="@+id/textView50"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM"
-                android:textAllCaps="true"
-
-
-                android:textColor="#fff"
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView51"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView52"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView53"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView54"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView56"
-                android:layout_width="200dp"
-                android:layout_height="104dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-            <TextView
-                android:id="@+id/textView49"
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_gravity="center"
-                android:text="ALARM!"
-                android:textAllCaps="true"
-                android:textColor="#fff"
-
-                android:textSize="50sp" />
-
-
-        </LinearLayout>
+            android:orientation="vertical" />
 
 
     </ScrollView>
-
-    <Button
-        android:id="@+id/cancelAlarmButton"
-        android:layout_width="173dp"
-        android:layout_height="50dp"
-        android:text="Button"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.421"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.083" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
