+++ app/src/main/res/values/styles.xml
@@ -8,4 +8,6 @@
         <item name="colorAccent">@color/colorAccent</item>
     </style>
 
+
+
 </resources>
